feat(deps)!: upgrade to baikai 0.3
Bump baikai, baikai-claude, baikai-openai, and baikai-effectful bounds
from >=0.2 && <0.3 to >=0.3 && <0.4 across all packages.

baikai 0.3 moves provider/registry/CLI failures in-band: completeRequest
no longer throws BaikaiError but returns an error-shaped Response whose
responseError is populated. The LLM interpreters relied on BE.complete
throwing, so an error response would have silently passed through as a
success and runLLMResilient would never retry it. Add raiseResponseError
(the blocking-path analogue of raiseStreamError) to both bareHandler and
runLLMResilient, inspecting responseError and throwing fromBaikaiError;
the resilient path charges the budget before raising, mirroring the
stream path.

Response.latencyMs changed Integer -> Int; fromIntegral it where it flows
into the Maybe Integer span attribute in Shikumi.Trace.

Adapt test/example fixtures to 0.3's event shapes: StartPayload gained a
responseId field and doneTerminal is now
Maybe Text -> StopReason -> Message -> TerminalPayload. baikai 0.3 also
exports flattenAssistantText, which collided with StubProvider's local
helper — hide baikai's from the open imports in StubProvider and LiveSpec.

BREAKING CHANGE: requires baikai 0.3.x.

+++ shikumi-trace/src/Shikumi/Trace.hs
@@ -342,7 +342,7 @@ llmAttrs m c o resp =
     & #provider ?~ (m ^. #provider)
     & #prompt ?~ requestToCanonicalValue m c o
     & #response ?~ toJSON resp
-    & #latencyMs ?~ (resp ^. #latencyMs)
+    & #latencyMs ?~ fromIntegral (resp ^. #latencyMs)
     & #inputTokens ?~ (resp ^. #message . #usage . #inputTokens)
     & #outputTokens ?~ (resp ^. #message . #usage . #outputTokens)
     & #costUsd ?~ realToFrac (resp ^. #message . #usage . #cost . #usd :: Rational)
